+++ encoding/tags.go
@@ -8,7 +8,10 @@ import (
 	"unicode"
 )
 
-const TagName = "gorethink"
+const (
+	TagName    = "gorethink"
+	RefTagName = "gorethink_ref"
+)
 
 // tagOptions is the string following a comma in a struct field's
 // tag, or the empty string. It does not include the leading comma.
@@ -16,6 +19,10 @@ type tagOptions string
 
 func getTag(sf reflect.StructField) string {
 	return sf.Tag.Get(TagName)
+}
+
+func getRefTag(sf reflect.StructField) string {
+	return sf.Tag.Get(RefTagName)
 }
 
 // parseTag splits a struct field's tag into its name and